move activation quant kernels to quant_utils.py
